Add files via upload
Build a dynamic online platform for travel enthusiasts, offering a seamless journey of exploration and
adventure. It provides an extensive database of destinations and curated travel packages with competitive pricing and
exclusive discounts. It accommodates solo travelers, couples, families, and groups, ensuring ease and convenience in
booking. It offers multiple communication channels and a special first-booking discount to enhance customer experience. It is a frontend project and I developed the front-end for the project using HTML and CS